// Client/src/app/services/menumodel.ts
export class Ditore {
    _id: string;
    food: string;
    description: string;

}

export class Drink {
    _id: string;
    drink: string;
    descriptiondr: string;

}

export class Dessert {
    _id: string;
    dessert: string;
    descriptiond: string;

}

export class Pasta {
    _id: string;
    pasta: string;
    descriptionp: string;

}

export class Meat {
    _id: string;
    meat: string;
    descriptionm: string;

}

export class Fish {
    _id: string;
    fish: string;
    descriptionf: string;

}
export class Admin {
    _id: string;

    username: string;
    password: string;

}